+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "slate"
-version = "0.0.6"
+version = "0.1.0"
 description = "A discord.py compatible wrapper for the lavalink/obsidian websocket and rest API."
 license = "MIT"
 authors = ["Axel <axelancerr@gmail.com>"]
@@ -48,7 +48,7 @@ sphinx =                { version = "^4.4.0",       optional = true }
 sphinxcontrib-trio =    { version = "^1.1.2",       optional = true }
 sphinx-copybutton =     { version = "^0.5.0",       optional = true }
 sphinxext-opengraph =   { version = "^0.5.1",       optional = true }
-furo =                  { version = "^2022.1.2",  optional = true }
+furo =                  { version = "^2022.1.2",    optional = true }
 
 
 [tool.poetry.dev-dependencies]
