+++ slate/objects/__init__.py
@@ -2,4 +2,10 @@
 from __future__ import annotations
 
 # My stuff
-from slate.objects.enums import *
+from .collection import *
+from .enums import *
+from .events import *
+from .filters import *
+from .result import *
+from .stats import *
+from .track import *
