+++ slate/objects/filters.py
