--- slate/obsidian/__init__.py
@@ -1,8 +1,0 @@
-# Future
-from __future__ import annotations
-
-# My stuff
-from slate.obsidian.exceptions import *
-from slate.obsidian.node import *
-from slate.obsidian.objects import *
-from slate.obsidian.player import *
