--- slate/obsidian/objects/__init__.py
@@ -1,11 +1,0 @@
-# Future
-from __future__ import annotations
-
-# My stuff
-from slate.obsidian.objects.collection import *
-from slate.obsidian.objects.enums import *
-from slate.obsidian.objects.events import *
-from slate.obsidian.objects.filters import *
-from slate.obsidian.objects.result import *
-from slate.obsidian.objects.stats import *
-from slate.obsidian.objects.track import *
